+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Fix: Array object type
 
-* If the items types is 'object' the array return undefined
+- If the items types is 'object' the array return undefined
 
 ## [1.0.2] (2023-08-26)
 
-* Initial version
+- Initial version